Add files via upload
This data structure files includes a self-SOAP reshape function which cast the SS power vector to the same shape of the IS power vector. This means that the ML model then needs to involve just the set of functions for the IS prediction, since these can be used to predict the SS as well.
This is useful for both the code and the model generalisation.